Adds missing include and avoids unused variable
This commit fixes two small issues:

IAMF_decoder_private.h uses IAMF_extradata and IAMF_StreamInfo but does
not include IAMF_decoder.h (And it does not appear to be transitively
included, so this commit adds the include.

In IAMF_decoder.c, if binauralizer is disabled, there is a unused var
`ctx`.  This PR moves it inside the code block where it's used to
avoid that error.