Letting dep_codecs/build.sh inherit C[XX]FLAGS content
CMAKE_C_FLAGS in cmake cache - if not pre-defined - is initialised using environment variables in combination with builtin defaults for the toolchain
If it is passed as an argument to the cmake configuration command, only the defined values will be cached.
Inheriting flags (e.g. optimisation options) from environment variables is not possible in this configuration
Passing the "-fPIC" flag however can also be done without affecting CFLAGS inheritance via setting the CMAKE_POSITION_INDEPENDENT_CODE cmake variable